get2gnow #0000DDaC4 - *tick-tock*...get ready, get set, go!
Alrighty about time, yes, I've had time to focus on working on get2gnow, starting with a 'from scratch' setup.  Yep, I nuked everything & did a fresh install, fixed this - tweaked that - rewrote another thing.  Than I did it all over again, and again, and again.

So I've gotten to the route of a ton problems & random issues.  Now for starters, new accounts can be created, deleted, connected to, & fiddled with pretty much all you want.  You may get gtk to toss some assert messages to the console but just try & crash it...I dare you ;-)

I'm not trying to come off as arrogant, only demonstrating that I've fixed a ton today.  Though that by no way means I can promise 'every things' fixed - just a crap load.

Several issues with creating new accounts have been fixed.  Some where some issues with the ServiceDialog buttons acting weird, others where causing incorrect disconnections, & just random weirdness.

I've also done a ton of work to finish my work on my new 'Connect' button & feature.  That said its not working all of the time, especially not with 'unsaved' accounts, so to avoid previous confusions, I've just 'hidden' the button for now, until I have it working under all possible situations, better to avoid confussion than to expect or even hope any of this ever get's ready by any one.

I also completely rewrote my primary gconfig method, namely, 'gconfig_if_bool'.  So it now correctly loads & sets default values even if the default scheme has not been installed yet.  This will mean a much faster start-up & lower memory foot print for any one who hasn't tweaked their preferences.

I've also completely refactored my background monitoring of DMs & @ replies & they're all working correctly again & much faster than before.  I'm going to be doing a complete reworking of how get2gnow processes timelines.  I've extended it long past its original implementation & now the code I 'inherited' is more 'in the way' & becoming stuff I have to hack around, so its time to take a scalpel to it & get it working as efficiantly as possible.

Lastly I've combined the old 'confirm' code into my 'PopupDialog' so its now all in one code base & working from one GtkBuilder file w/ver lil 'manual' gtk code creating its UI.

BTW, if you run in to any bugs, after I've spent 13hrs straight tracking every thing I can down, than I'm sorry & I'd really appriciate it if you could send me your 'bug buddy' 'bugreport.txt' file.

I've said that because, well there's a reason all software's 'AS IS', just cause I can promise.  It doesn't mean I haven't tried & doesn't mean I don't care.  I'm just doing this on my own & doing all I can, so please, if you do find a bug, be polite, be nice, & I'll care.  Be rude, blunt, or an ass & well...my suggestion is to use a cactus as colonic.  Its even natural.

That let out, I'm sorry having a slightly 'busted' version of get2gnow listed as stable for so long, not because I could have done any thing about it.  I'm only sorry if any one thought I could have.  After all when I say I'm a disabled OSS artist, its not an expression, its the reason I can no longer work as a programmer.  I have no control or knowledge over what I will or won't be able to do from one day to another.  Hell I don't have knowledge or control over any of it from one moment to the next...so its something we all just have to deal with & get over.

One other big fix/change I've finally 'nailed down' is that favoriting tweets works again.  But far more important is that I've finally fixed get2gnow so it now correctly posts replies as 'reply to...' not only on twitter but now on Identi.ca & all Laconica instances as well.

Lastly I've fixed all of the direct messaging methods & I've added a key board shortcut, 'Shift+S'.  Which if you have an updated selected than it will send your update/tweet as a direct message to the user who wrote the tweet you've selected.

So yeah, after days of being 'down & out' I've got a ton done.  I'll keep working to make get2gnow the best I can, while I can, until I go done again, or go down for the last.  I'm really excited & hopeful for releasing get2gnow & even more for whatever's next...I'm happy to hope I'll have something next.

As always any feed back would be deeply & truly appriciated, more than I can express in a git commit log.  But I'd be thankful for any feedback, thanks, ideas, help, bugreports, & even just a 'hello'.

Committed on: 2009-06-07 @ 02:31:02 AM

+++ src/parser.c
@@ -430,7 +430,7 @@ void parser_format_user_status(OnlineService *service, User *user, UserStatus *s
 	g_free(sexy_status_text);
 }/*parser_format_user_status(status, user);*/
 
-gchar *parser_convert_time(const gchar *datetime, guint *my_diff){
+gchar *parser_convert_time(const gchar *datetime, gulong *my_diff){
 	struct tm	*ta;
 	struct tm	post;
 	int			 seconds_local;
