+++ src/parser.h
@@ -38,7 +38,7 @@ xmlDoc *parser_parse(SoupMessage *xml, xmlNode **first_element);
 
 void parser_format_user_status(OnlineService *service, User *user, UserStatus *status);
 gchar *parser_escape_text(gchar *status);
-gchar *parser_convert_time(const gchar *datetime, guint *my_diff);
+gchar *parser_convert_time(const gchar *datetime, gulong *my_diff);
 
 guint parser_timeline(OnlineService *service, SoupMessage *xml, StatusMonitor monitoring);
 gchar *parser_parse_xpath_content(SoupMessage *xml, const gchar *xpath);
